requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
rapidfuzz>=3.0.0
streamlit>=1.33.0
openai>=1.30.0
python-dotenv>=1.0.0
